refactor(test): rename basic test case file
