+++ test/test_g2p_cli.py
@@ -5,8 +5,8 @@ import os
 import tempfile
 from unittest import TestCase, main
 
+from basic_test_case import BasicTestCase
 from lxml import etree
-from utils import BasicTestCase
 
 from readalongs.align import align_audio
 from readalongs.app import app
