+++ test/test_prepare_cli.py
@@ -8,7 +8,7 @@ import tempfile
 from shutil import copyfile
 from unittest import TestCase, main
 
-from utils import BasicTestCase
+from basic_test_case import BasicTestCase
 
 from readalongs.align import create_input_tei
 from readalongs.app import app
